pyproject.toml
[tool.poetry]
name = "simple-chat"
version = "0.1.0"
description = "Test task"
authors = ["Ivan Grakov"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
django = "==5.0"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
factory-boy = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
